worked on controllers and services and foreign key mapping in the database

+++ Spring/src/main/java/com/revature/Group4P2/beans/controllers/CartControllers.java
@@ -1,12 +1,19 @@
 package com.revature.Group4P2.beans.controllers;
 
 import com.revature.Group4P2.beans.services.CartService;
+import com.revature.Group4P2.beans.services.CatalogService;
+import com.revature.Group4P2.beans.services.UserService;
 import com.revature.Group4P2.entities.Cart;
+import com.revature.Group4P2.entities.Catalog;
+import com.revature.Group4P2.entities.CatalogDetails;
+import com.revature.Group4P2.entities.Users;
+import org.apache.catalina.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
+import javax.swing.text.html.Option;
 import java.util.List;
 import java.util.Optional;
 
@@ -15,10 +22,15 @@ import java.util.Optional;
 public class CartControllers {
     private CartService service;
 
+    // catalog and user services
+    private CatalogService catalogService;
+    private UserService userService;
 
     @Autowired
-    public CartControllers(CartService service) {
+    public CartControllers(CartService service, CatalogService catalogService, UserService userService) {
         this.service = service;
+        this.catalogService = catalogService;
+        this.userService = userService;
     }
 
     //5 crud things:
@@ -43,7 +55,20 @@ public class CartControllers {
     @ResponseStatus(value = HttpStatus.OK)
     public void createCart(@RequestBody Cart cart)
     {
-        service.createCart(cart);
+        Optional<Catalog> optionalCatalog = catalogService.getCatalogById(cart.getCatalog().getItemId());
+        Optional<Users> optionalUsers = userService.getUserById(cart.getUser().getId());
+
+        if(optionalCatalog.isPresent() && optionalUsers.isPresent()) {
+            cart.setCatalog(catalogService.getCatalogById(cart.getCatalog().getItemId()).get());
+            cart.setUser(userService.getUserById(cart.getUser().getId()).get());
+            service.createCart(cart);
+        }
+        else
+        {
+            // throw exception later on
+            System.out.println("CART ITEM NOT CREATED");
+        }
+
     }
 
 
@@ -59,7 +84,16 @@ public class CartControllers {
     @ResponseStatus(value = HttpStatus.ACCEPTED)
     public void deleteCart(@PathVariable Integer cartId)
     {
-        service.deleteCartById(cartId);
+        Optional<Cart> optionalCart = service.getCartById(cartId);
+        if(optionalCart.isPresent())
+        {
+            Cart cart = optionalCart.get();
+            cart.setUser(null);
+            cart.setCatalog(null);
+            service.updateCart(cart);
+            service.deleteCartById(cartId);
+        }
+
     }
 
 }
