+++ Spring/src/main/java/com/revature/Group4P2/beans/controllers/CatalogControllers.java
@@ -11,6 +11,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
+import javax.swing.text.html.Option;
 import java.util.List;
 import java.util.Optional;
 
@@ -65,13 +66,29 @@ public class CatalogControllers {
     @ResponseStatus(value = HttpStatus.ACCEPTED)
     public void updateCatalog(@RequestBody Catalog catalog)
     {
+
+        System.out.println("UPDATE");
         service.updateCatalog(catalog);
+        System.out.println("UPDATING");
     }
     // DELETE - delete - delete
     @RequestMapping(value = "/{catId}", method = RequestMethod.DELETE)
     @ResponseStatus(value = HttpStatus.ACCEPTED)
     public void deleteCatalog(@PathVariable Integer catId)
     {
-        service.deleteCatalogById(catId);
+        // need to get the foreign key references and dereference them
+        Optional<Catalog> optionalCatalog = service.getCatalogById(catId);
+        if(optionalCatalog.isPresent())
+        {
+            Catalog catalog = optionalCatalog.get();
+            // update the thing in the data base with null foreign keys
+            catalog.setCatalogDetails(null);
+            service.updateCatalog(catalog);
+            service.deleteCatalogById(catId);
+        }
+        else {
+            // throw exception
+        }
+
     }
 }
