+++ Spring/src/main/java/com/revature/Group4P2/beans/controllers/PurchasesController.java
@@ -1,14 +1,19 @@
 package com.revature.Group4P2.beans.controllers;
 
 import com.revature.Group4P2.beans.services.CartService;
+import com.revature.Group4P2.beans.services.CatalogService;
 import com.revature.Group4P2.beans.services.PurchaseService;
+import com.revature.Group4P2.beans.services.UserService;
 import com.revature.Group4P2.entities.Cart;
+import com.revature.Group4P2.entities.Catalog;
 import com.revature.Group4P2.entities.Purchases;
+import com.revature.Group4P2.entities.Users;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
+import javax.swing.text.html.Option;
 import java.util.List;
 import java.util.Optional;
 
@@ -16,9 +21,16 @@ import java.util.Optional;
 @RequestMapping(value = "/purchases")
 public class PurchasesController {
     private PurchaseService service;
+    private CatalogService catalogService;
+    private UserService userService;
+
+    private  CartService cartService;
     @Autowired
-    public PurchasesController(PurchaseService service) {
+    public PurchasesController(PurchaseService service, CatalogService catalogService, UserService userService, CartService cartService) {
         this.service = service;
+        this.catalogService = catalogService;
+        this.userService = userService;
+        this.cartService = cartService;
     }
 
     //5 crud things:
@@ -44,7 +56,27 @@ public class PurchasesController {
     @ResponseStatus(value = HttpStatus.OK)
     public void createPurchases(@RequestBody Purchases purchases)
     {
-        service.createPurchases(purchases);
+        // catalog and user
+        Optional<Catalog> optionalCatalog = catalogService.getCatalogById(purchases.getCatalog().getItemId());
+        Optional<Users> optionalUsers = userService.getUserById(purchases.getUser().getId());
+        Optional<Cart> optionalCart = cartService.getCartById(purchases.getCart().getItemId());
+        if(optionalUsers.isPresent() && optionalCatalog.isPresent() && optionalCart.isPresent())
+        {
+            purchases.setCatalog(catalogService.getCatalogById(purchases.getCatalog().getItemId()).get());
+            purchases.setUser(userService.getUserById(purchases.getUser().getId()).get());
+            purchases.setCart(cartService.getCartById(purchases.getCart().getItemId()).get());
+            purchases.setQuantityPurchased(cartService.getCartById(purchases.getCart().getItemId()).get().getQuantity());
+            Double totalAmount = purchases.getCart().getCatalog().getItemPrice() * purchases.getQuantityPurchased();
+            purchases.setTotalAmount(totalAmount);
+            service.createPurchases(purchases);
+        }
+
+        else
+        {
+            // throw exception later on
+            System.out.println("Purchase ITEM NOT CREATED");
+        }
+
     }
 
 
@@ -60,6 +92,16 @@ public class PurchasesController {
     @ResponseStatus(value = HttpStatus.ACCEPTED)
     public void deletePurchases(@PathVariable Integer purchaseId)
     {
-        service.deletePurchasesById(purchaseId);
+        Optional<Purchases> optionalPurchases = service.getPurchaseById(purchaseId);
+        if(optionalPurchases.isPresent())
+        {
+            // then set its foreign keys to null and update the db
+            Purchases purchases = optionalPurchases.get();
+            purchases.setUser(null);
+            purchases.setCatalog(null);
+            service.updatePurchases(purchases);
+            service.deletePurchasesById(purchaseId);
+        }
+
     }
 }
